pyproject.toml
[project]
name = "invoice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]
